Updated the website_scraping function and now the return value is put to a .txt file. Also cleaned up code by dividing it into different files.
